webapp/pyproject.toml
[project]
name = "brico"
version = "0.1.0"
description = ""
authors = [
    { name = "Nicolas Oudard", email = "nicolas@oudard.org" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django-allauth>=65.4.0",
    "django-cors-headers>=4.7.0",
    "django>=5.1.4",
    "djangorestframework>=3.14.0",
    "environs>=14.1.0",
    "honcho>=2.0.0",
    "requests>=2.32.3",
    "psycopg>=3.2.12",
    "dj-rest-auth>=7.0.1",
    "pytest>=8.4.2",
    "pytest-django>=4.9.0",
    "factory-boy>=3.3.0",
    "pydantic-ai>=1.16.0",
    "pydantic-ai-examples>=1.16.0",
    "django-import-export[xls,xlsx]>=4.3.14",
    "django-extensions>=4.1",
    "markdown>=3.7",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.9.5",
    "pre-commit>=4.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["audits", "core"]


[tool.ruff.lint]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

per-file-ignores = {}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "ˆ(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
